Actually, use DB_NAME=test_uploads for testing
The primary HSR test job `hsreplaynet-web-test` maintains a set of Postgres
databases on build.hearthsim.net for the purposes of local integration
testing. The deploy job for the isolated Lambda already depends on them,
which is a little bit unhygienic; this change simply updates that
dependency to match the current behavior of `hsreplaynet-web-test`.